added inventory layer

+++ services/products-service/domain/inventory.go
@@ -9,7 +9,7 @@ import (
 type Inventory struct {
 	ID        primitive.ObjectID `bson:"_id,omitempty"`
 	ProductID primitive.ObjectID `bson:"product_id"`
-	Quantity  int64              `bson:"quantity"`
+	Quantity  uint32             `bson:"quantity"`
 	CreatedAt time.Time          `bson:"created_at"`
 	UpdatedAt time.Time          `bson:"updated_at"`
 }
